+++ docs/releases.md
@@ -6,7 +6,7 @@ This page list all of the changes that came during the different release of the 
 
 * adding `FieldGroupManager` and `SchemaManager` class in the `schema` submodule
 * adding support for more parameter on `getFailedBatches` in `catalog` submodule
-* 
+* adding the `FlowManager` class to gather all information from one flow Id.
 
 ## version 0.2.1